feat : Save all feature is deleted, and snackbar has 'next feature'
Since the user will inserts stamp and saves it, the 'Save All' feature
is deleted because it will cause only potential bug.
And if the user saves the preview, a snackbar is displayed to indicate
its location. And next If user press "NEXT" button on that snackbar,
it will show next preview to user.